pyproject.toml
[tool.poetry]
name = "streamlit-tools"
version = "0.1.0"
description = "Streamlit application tools"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
streamlit = "1.29.0"
pandas = "2.1.4"
numpy = "1.26.2"
streamlit-authenticator = "0.2.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
